refactor: FeaturedFairs section
